requirements.txt
asgiref==3.3.1
boto3==1.17.21
botocore==1.20.21
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
configparser==5.0.2
Django==3.1.7
django-storages==1.11.1
djangorestframework==3.12.2
gunicorn==20.0.4
idna==2.10
jmespath==0.10.0
Pillow==8.1.1
psycopg2-binary==2.8.6
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
s3transfer==0.3.4
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.3
